derp discovered bug: Pie Chart logic wrong :pill::beers:
