+++ src/containers/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Chart from './Chart';
+import ChargeList from './ChargeList';
 // import AddCharge from './AddCharge';
-import Pie from './PieChart';
-import Bar from './BarChart';
+import PieChart from './PieChart';
+import BarChart from './BarChart';
 import Categories from './Categories';
 import FileInput from '../components/FileInput';
 
@@ -11,10 +11,10 @@ const App = () => {
     <div className="col-12 clearfix">
       <div className="wrapper col-8 mx-auto">
         <h1 className="blue pl2">TD VISA READER !!!!!!!!</h1>
-        <Chart />
+        <ChargeList />
         <FileInput />
-        <Bar />
-        <Pie />
+        <BarChart />
+        <PieChart />
         { /* <AddCharge /> */ }
         <Categories />
       </div>
